refactor: getMarkdownPreviewProps has same name patter as json equivalent
